+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Changed minimum supported rust version from 1.57 to 1.60.
+
 ### Fixed
 - Fixed a typo in the `Exti::set_falling_trigger` function name.