+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.0] - 2022-05-08
 ### Added
 - Added `set_sleep_clock` to GPIO ports to enable and disable clocks during sleep.
 - Added `subghz::Timeout::saturating_add`.
@@ -94,7 +94,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [tweedegolf]: https://github.com/tweedegolf
 [stm32-rs]: https://github.com/stm32-rs
-[Unreleased]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.5.0...HEAD
+[0.4.1]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.2.1...v0.3.0